chore: move custom generator to the original file
